I created an enumerator script for lootable objects. Idle, Opening, Open, Closing.
I added some additional sprites to my sLootbag thus oLootableItem has a possible animation.

I managed to get the Left Click event to change the state of said oLootableItem.

Then, if it is in the opened state && the player is within 20 pixels that can loot it!  By loot I mean, grab the item and destroy the oLootableItem.
^ not perfect, but it successfully loots the opened bag and it poofs with the item added to inventory.

Also, the player can click on a opened oLootableItem (the bag) if they are within 20 pixels, and close it (as long as they do not run into it... causing it to be destroyed).
^ again not perfect, but it is something!